feat: robots.txt as page
